requirements.txt
Adafruit-Blinka==3.0.0
adafruit-circuitpython-neopixel==3.4.0
Adafruit-PlatformDetect==1.3.4
Adafruit-PureIO==0.2.3
board==1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
DateTime==4.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycairo==1.18.2
pygame==1.9.6
PyGObject==3.34.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
spidev==3.4
sysv-ipc==1.0.0
uritemplate==3.0.0
urllib3==1.25.6
zope.interface==4.6.0
